Refactor javascripts to match the new layouts.
Added content_editor and accounts layouts. This commit
refactors the javascripts for those layouts to use properly.

Note: Rails 6 moved from the Sprockets assets pipeline to a Webpack
pipeline for javascripts. This removes the old Sprokects javascripts files
which weren't used. See: https://prathamesh.tech/2019/08/26/understanding-webpacker-in-rails-6/

Test Plan:
- load the sign-in and sign-up pages and make sure they link in the correct javascripts for that layout.
- same for the content_editor.
