+++ app/javascript/packs/accounts.js
@@ -1,0 +1,5 @@
+// This is the main entrypoint for the accounts layout
+// Things that need to be set up globally for all accounts related pages go here.
+// Run this script by adding <%= javascript_pack_tag 'accounts' %> to the head of your layout file,
+// like app/views/layouts/accounts.html.erb.
+
