+++ app/javascript/packs/application.js
@@ -7,8 +7,15 @@
 // To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
 // layout file, like app/views/layouts/application.html.erb
 
-import Rails from '@rails/ujs';
-Rails.start();
+
+// Old assets pipeline (Sprockets) stuff that we're moving to Webpacker. These used to 
+// be in in: app/assets/javascripts/application.js
+require("@rails/ujs").start()
+require("turbolinks").start() // Note: this has to be run before any WebpackerReact.setup()
+// TODO: if we need to use either of these, start from a fresh Rails 6 install to figure out
+// what's missing that causes:  "Module not found: Error"
+//require("@rails/activestorage").start()
+//require("channels")
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
@@ -17,8 +24,21 @@ Rails.start();
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
 
-console.log('Hello World from Webpacker')
+// Uncomment and update the context directory to where you want to load all React components 
+// that are applicable to the entire application.
+// react-rails
 // Support component names relative to this directory:
-var componentRequireContext = require.context("components", true);
-var ReactRailsUJS = require("react_ujs");
-ReactRailsUJS.useContext(componentRequireContext);
+//
+//var componentRequireContext = require.context("app_components", true);
+//var ReactRailsUJS = require("react_ujs");
+//ReactRailsUJS.useContext(componentRequireContext);
+
+// Use axe a11y testing in development.
+import React from 'react'
+import ReactDOM from 'react-dom'
+if (process.env.NODE_ENV !== 'production') {
+  let axe = require('react-axe');
+  document.addEventListener('DOMContentLoaded', () => {
+    axe(React, ReactDOM, 1000);
+  });
+}
