+++ app/javascript/packs/server_rendering.js
@@ -1,3 +1,4 @@
+// react-rails
 // By default, this pack is loaded for server-side rendering.
 // It must expose react_ujs as `ReactRailsUJS` and prepare a require context.
 var componentRequireContext = require.context("components", true);
