+++ app/views/course_contents/_form.html.erb
@@ -1,6 +1,3 @@
-<%= javascript_pack_tag 'sidebar' %>
-<%= javascript_pack_tag 'content_editor' %>
-
 <%= react_component('Sidebar', {
     courseContents: @course_contents
 }) %>
