+++ app/views/home/welcome.html.erb
@@ -1,8 +1,6 @@
 <% content_for :headline do %>
   Dashboard
 <% end %>
-
-<%= react_component('Hello', name: 'React') %>
 <br/>
 <div>
 <%= link_to 'Content Editor', new_course_content_path %>
