+++ app/views/layouts/accounts.html.erb
@@ -1,7 +1,8 @@
 <% content_for :title, 'Braven Portal'  %>
 
 <% content_for :head do %>
-  <%= stylesheet_link_tag 'accounts' %>
+  <%= stylesheet_link_tag 'accounts', media: 'all', 'data-turbolinks-track': 'reload' %>
+  <%= javascript_pack_tag 'accounts' %>
 <% end %>
 
 <% content_for :content do %>
