+++ app/views/layouts/application.html.erb
@@ -10,8 +10,7 @@
 
     <%= braven_stylesheet_link_tag %>
     <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_pack_tag 'platform' %>
+    <%= javascript_pack_tag 'application' %>
 
     <%= yield :head %>
   </head>
