+++ app/views/layouts/content_editor.html.erb
@@ -1,5 +1,11 @@
 <% content_for :title, 'Content Editor' %>
 
+<% content_for :head do %>
+    <%= stylesheet_link_tag 'content_editor', media: 'all', 'data-turbolinks-track': 'reload' %>
+    <%= javascript_pack_tag 'sidebar' %>
+    <%= javascript_pack_tag 'content_editor' %>
+<% end %>
+
 <% content_for :content do %>
   <%# Add stuff to the body here %>
   <%= yield %>
